make source and dist directories
