+++ src/snake.h
@@ -2,6 +2,7 @@
 #define snake_h
 
 #include <unistd.h>
+#include <curses.h>
 #include <stdlib.h>
 #include <time.h>
 #include <ctype.h>
@@ -13,10 +14,13 @@ typedef struct {
     Position* tail;
 } Snake;
 
-void initSnake(int my);
+void initSnake(Snake* snake, int my);
+Position* bottomPos();
 void startGame();
 void gameOver();
 void moveHead(Snake* snake, int dir);
-void moveTail(Position* tail);
+void moveTail(Snake* snake, int mx, int my);
+void freeCorners();
+void printDebug(Snake* snake, int mx, int my);
 
 #endif
